+++ apps/server/cypress/e2e/suite1/search.cy.ts
@@ -36,7 +36,7 @@ describe("search", () => {
     ).should("exist");
 
     cy.visit(
-      `/widget?jobTypes=${ComboJobTypes.TRANSACTIONS},${ComboJobTypes.ACCOUNT_NUMBER},${ComboJobTypes.ACCOUNT_OWNER}&userId=${crypto.randomUUID()}`,
+      `/widget?jobTypes=${ComboJobTypes.TRANSACTIONS},${ComboJobTypes.ACCOUNT_NUMBER},${ComboJobTypes.ACCOUNT_OWNER}&userId=${crypto.randomUUID()}&targetOrigin=http://localhost:8080`,
     );
 
     cy.findByText(institutionThatIsInFavoriteAndSupportsAll).should("exist");