+++ apps/server/src/widgetEndpoint.ts
@@ -28,7 +28,9 @@ export const widgetHandler = (req: Request, res: Response) => {
     userId: Joi.string().required(),
     token: Joi.string(),
     aggregatorOverride: Joi.string().valid(...nonTestAggregators),
-    targetOrigin: Joi.string(),
+    targetOrigin: Joi.string()
+      .uri({ scheme: ["http", "https"] })
+      .required(),
   })
     .with("aggregator", ["institutionId", "connectionId"])
     .with("connectionId", ["institutionId", "aggregator"]);