+++ packages/finicity-adapter/playwright/finicity.playwright.ts
@@ -13,7 +13,7 @@ const makeAConnection = async (
   userId: string,
 ) => {
   await page.goto(
-    `http://localhost:8080/widget?jobTypes=${jobTypes.join(",")}&userId=${userId}`,
+    `http://localhost:8080/widget?jobTypes=${jobTypes.join(",")}&userId=${userId}&targetOrigin=http://localhost:8080`,
   );
 
   page.evaluate(`
@@ -169,7 +169,7 @@ test.describe("Finicity Adapter Tests", () => {
     await testDataEndpoints({ request, userId, connectionId, aggregator });
 
     await page.goto(
-      `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}&aggregator=${aggregator}&institutionId=${ucpInstitutionId}&connectionId=${connectionId}`,
+      `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}&aggregator=${aggregator}&institutionId=${ucpInstitutionId}&connectionId=${connectionId}&targetOrigin=http://localhost:8080`,
     );
 
     await expect(page.getByText("Log in at FinBank Profiles - A")).toBeVisible({
@@ -195,7 +195,7 @@ test.describe("Finicity Adapter Tests", () => {
     test.setTimeout(300000);
 
     await page.goto(
-      `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}`,
+      `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}&targetOrigin=http://localhost:8080`,
     );
 
     await page.getByPlaceholder("Search").fill("finbank");