+++ packages/utils-e2e/cypress/visit.ts
@@ -10,7 +10,9 @@ export const visitWithPostMessageSpy = (url: string) =>
 export const visitIdentity = () => {
   const userId = crypto.randomUUID();
 
-  cy.visit(`/widget?jobTypes=${ComboJobTypes.ACCOUNT_OWNER}&userId=${userId}`);
+  cy.visit(
+    `/widget?jobTypes=${ComboJobTypes.ACCOUNT_OWNER}&userId=${userId}&targetOrigin=http://localhost:8080`,
+  );
 
   return cy.wrap(userId);
 };
@@ -32,7 +34,7 @@ export const visitAgg = (options?: any) => {
     : "";
 
   cy.visit(
-    `/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}${tokenString}${aggregatorOverrideString}`,
+    `/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}${tokenString}${aggregatorOverrideString}&targetOrigin=http://localhost:8080`,
     {
       failOnStatusCode,
     },